+++ submodules/TelegramCore/Sources/TelegramEngine/Peers/PeerPhotoUpdater.swift
@@ -171,7 +171,7 @@ func _internal_updatePeerPhotoInternal(postbox: Postbox, network: Network, state
                                                             switch size {
                                                                 case let .videoSize(_, type, w, h, size, videoStartTs):
                                                                     let resource: TelegramMediaResource
-                                                                    resource = CloudPhotoSizeMediaResource(datacenterId: dcId, photoId: id, accessHash: accessHash, sizeSpec: type, size: Int(size), fileReference: fileReference.makeData())
+                                                                    resource = CloudPhotoSizeMediaResource(datacenterId: dcId, photoId: id, accessHash: accessHash, sizeSpec: type, size: Int64(size), fileReference: fileReference.makeData())
                                                                     
                                                                     videoRepresentations.append(TelegramMediaImage.VideoRepresentation(dimensions: PixelDimensions(width: w, height: h), resource: resource, startTimestamp: videoStartTs))
                                                             }
